code split Hero Section
